Remove most construction arguments for DspDsmKsPDF
and make proper defaults which can be modified by the commandline. Helps
to remove the number of places we need to define the analysis window and
intended to calculate the efficiency more correctly with a small
program that just uses DspDsmKsPDF and loads the data set.

Also the quality cuts are now optional to remove them for toymc
generation